Setup EBX Connection

This function will store the SOAP connection details in a yaml file

SetupEBXConnection(meta_branch = "Fishery", meta_instance = "Fishery",
  ebx_soap_url, username, password)

Arguments

meta_branch
metadata branch name in EBX, used for faostatR library (optional)

meta_instance
metadata instance name in EBX, used for faostatR library (optional)

ebx_soap_url
SOAP webservice url

username
username, a character.

password
password, a character.

Value

Status 0 (zero) ok.

Details

In case of an shiny app, run the SetupEBXConnection from your Rstudio. The conncetion details are persisted to a file [ebx5_connecion.yml] which you can then embed into the shiny app. The library faoebx5 uses GetEBXConnection, in order to obtain the connection details from the yaml file. Obviously, the yaml file is excluded, not part of the GIT repository faoebx5.

See also

Author

Thomas Berger, thomas.berger@fao.org

Luis G. Silva e Silva, luis.silvaesilva@fao.org